adding project files and README
