+++ 03-GettingStarted/samples/javascript/index.js
