+++ 03-GettingStarted/samples/typescript/src/index.ts
