+++ 04-PracticalImplementation/samples/javascript/README.md
@@ -1,0 +1,17 @@
+# Sample
+
+This is a JavaScript sample for an MCP Server
+
+## Install
+
+Run the following command:
+
+```bash
+npm install
+```
+
+## Run
+
+```bash
+npm start
+```
