+++ 04-PracticalImplementation/samples/typescript/README.md
@@ -1,0 +1,17 @@
+# Sample
+
+This is a Typescript sample for an MCP Server
+
+## Install
+
+Run the following command:
+
+```bash
+npm install
+```
+
+## Run
+
+```bash
+npm start
+```
